refactor: use abstract class instead of CommandI
