+++ src/Command.ts
@@ -1,8 +1,19 @@
-import { ApplicationCommandOptionData } from 'discord.js';
-import { CommandOptions } from './typings';
-import { CommandType } from './typings/enums';
+import { ApplicationCommandOptionData, ChatInputCommandInteraction } from 'discord.js';
+import { Bot } from './Bot';
 
-export class Command {
+export enum CommandType {
+    Help, Image, Misc
+}
+
+export interface CommandOptions {
+    args?: ApplicationCommandOptionData[]
+    help: string
+    name: string
+    usage: string
+    type: CommandType
+}
+
+export abstract class Command {
     public args: ApplicationCommandOptionData[] | undefined;
     public name: string;
     public help: string;
@@ -16,4 +27,6 @@ export class Command {
         this.usage = usage;
         this.type = type;
     }
+
+    public abstract execute(client: Bot, interaction: ChatInputCommandInteraction): void
 }
