+++ src/commands/triggered.ts
@@ -1,11 +1,9 @@
 import { Canvacord } from 'canvacord';
 import { ApplicationCommandOptionType, AttachmentBuilder, ChatInputCommandInteraction } from 'discord.js';
 import { Bot } from '../Bot';
-import { Command } from '../Command';
-import { CommandInterface } from '../typings';
-import { CommandType } from '../typings/enums';
+import { Command, CommandType } from '../Command';
 
-export default class TriggeredCommand extends Command implements CommandInterface {
+export default class TriggeredCommand extends Command {
     constructor() {
         super({
             args: [{
