--- src/typings/enums.ts
@@ -1,3 +1,0 @@
-export enum CommandType {
-    Help, Image, Misc
-}
